# openapi.yaml
openapi: 3.1.0
info:
  title: msbd — Microsandbox REST host
  version: 0.6.0
  description: >
    A standalone, self-hostable HTTP server that wraps the microsandbox Go SDK
    (cgo + libkrun) and exposes a small REST API for managing fast, local
    microVMs. Run it on a host that has /dev/kvm; clients then treat
    microsandbox as a remote sandbox backend with no cgo on their side.


    Every versioned endpoint lives under `/api/v1`. The unversioned paths
    `/healthz`, `/readyz`, `/metrics`, `/docs` and `/openapi.yaml` sit at the
    root, and the root itself (`/`, `/sandboxes`, `/volumes`, `/images`,
    `/snapshots`, `/settings/*`) is the optional web dashboard, which
    authenticates with its own login session rather than a bearer token.


    All endpoints except `/healthz` and `/readyz` require a bearer token.


    Tokens come from either `--api-key` / `MSBD_API_KEY` (comma-separated for
    rotation) or msbd's own key store, managed with `msbd keys create <name>`
    and from the dashboard. Both kinds are accepted together; a stored key can
    be revoked (`msbd keys revoke`) or given an expiry, after which it returns
    401 like any other invalid token. If no key exists at all the server is
    unauthenticated — intended for local development only.
servers:
  - url: http://localhost:8099
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/terminal-tickets:
    post:
      tags: [meta]
      summary: Mint a short-lived single-use terminal ticket
      description: >
        Returns an opaque, ~60s, single-use ticket bound to one sandbox that the
        terminal WebSocket accepts as `?ticket=` instead of the long-lived
        bearer key — so browser clients never place the API key in a URL, page
        source, or proxy log. Minting requires normal bearer auth.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sandbox]
              properties:
                sandbox: { type: string, description: Sandbox id the ticket is valid for. }
      responses:
        "201":
          description: Ticket minted
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expires_at]
                properties:
                  ticket: { type: string }
                  expires_at: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/sandboxes:
    post:
      tags: [lifecycle]
      summary: Create (boot) a sandbox
      description: >
        Boots a detached microVM from an OCI image. Cold first pull of a tag can
        take minutes; cached boots are ~150ms. Returns the live instance.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201":
          description: Sandbox created and running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "400":
          description: Invalid parameters (out-of-range cpu/memory/ports, bad body)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "507":
          description: Host at capacity (MSBD_MAX_SANDBOXES reached). Boot/validation failures are 400/500, not 507.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
    get:
      tags: [lifecycle]
      summary: List all sandboxes
      responses:
        "200":
          description: All known sandboxes (running or stopped)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/sandboxes/{id}:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    get:
      tags: [lifecycle]
      summary: Get one sandbox
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [lifecycle]
      summary: Delete (tear down) a sandbox
      description: Stops the box if running, then removes it permanently.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/stop:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [lifecycle]
      summary: Stop (pause) a sandbox
      description: >
        Pauses the VM (releases CPU/mem). Does not auto-resume; the next
        run/launch/file call transparently resumes it.
      responses:
        "204": { description: Stopped }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/start:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [lifecycle]
      summary: Start (ensure running) a sandbox
      description: >
        Idempotent ensure-running; boots a stopped box and waits until its guest
        agent is actually reachable before returning. If the runtime reports a
        start but the guest never comes up (e.g. a halted guest that left stale
        runtime state), msbd force-kills and retries once, then fails with 500
        rather than reporting a false-positive 204 — such a sandbox is
        unresumable and must be recreated.
      responses:
        "204": { description: Running (guest agent verified reachable) }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500":
          description: >
            The sandbox reported a start but its guest agent never came up; it is
            unresumable and must be deleted and recreated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
  /api/v1/sandboxes/{id}/exec:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [exec]
      summary: Run a short command (one round-trip)
      description: For quick provisioning helpers (mkdir, chmod, pwd). Does not ensure-running.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecRequest" }
      responses:
        "200":
          description: Command finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/run:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [exec]
      summary: Run a long command (blocking, ensure-running)
      description: >
        Safe for long commands (builds, installs, test suites). Transparently
        resumes a paused box first. No server-imposed request ceiling — front
        with a proxy that has a high read timeout, or none.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecRequest" }
      responses:
        "200":
          description: Command finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/terminal:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    get:
      tags: [exec]
      summary: Interactive terminal (WebSocket upgrade)
      description: >
        Upgrades the connection to a WebSocket that bridges a terminal emulator
        (e.g. xterm.js) to a shell running inside the guest. This is NOT a JSON
        endpoint; it returns `101 Switching Protocols`.


        Wire protocol — client → server: a **binary** frame carries raw stdin
        bytes (keystrokes); a **text** frame carries a JSON control message,
        either `{"type":"resize","rows":40,"cols":120}` or
        `{"type":"signal","signal":2}` (2 = SIGINT). Server → client: a
        **binary** frame carries raw terminal output; a **text** frame carries
        a JSON event, currently only `{"type":"exit","exit_code":0}` sent just
        before the socket closes.


        Auth: send the bearer token via the `Authorization` header, or — for
        browser clients that cannot set headers on a WebSocket handshake — as
        the `key` query parameter.


        The shell runs on a real kernel PTY in the guest, so colors, line
        editing, window resize (via the `resize` control frame), job control,
        and full-screen TUIs (vim, top) all work.
      parameters:
        - name: cmd
          in: query
          required: false
          schema: { type: string }
          description: Command line to run. Empty launches `/bin/sh -i`; non-empty runs via `/bin/sh -c`.
        - name: cwd
          in: query
          required: false
          schema: { type: string }
          description: Working directory for the session.
        - name: rows
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 65535 }
          description: Initial PTY rows (default 24).
        - name: cols
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 65535 }
          description: Initial PTY columns (default 80).
        - name: term
          in: query
          required: false
          schema: { type: string, default: "xterm-256color" }
          description: $TERM value to export in the session.
        - name: key
          in: query
          required: false
          schema: { type: string }
          description: Bearer token, for WebSocket clients that cannot set the Authorization header.
      responses:
        "101":
          description: Switching Protocols — WebSocket established
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/jobs:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [jobs]
      summary: Launch a background command
      description: >
        Starts a command in the background and returns a job handle immediately.
        Output is captured as it streams; poll for state and accumulated output.
        Jobs are in-memory: after an msbd restart a prior job polls as "gone".
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecRequest" }
      responses:
        "202":
          description: Job launched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LaunchResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/jobs/{job}:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
      - name: job
        in: path
        required: true
        schema: { type: string }
        description: Job handle from the launch response.
    get:
      tags: [jobs]
      summary: Poll a background job
      responses:
        "200":
          description: Job status + output captured so far
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/sandboxes/{id}/files/read:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [files]
      summary: Read a file (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileReadRequest" }
      responses:
        "200":
          description: File contents, base64-encoded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileReadResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/write:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    post:
      tags: [files]
      summary: Write a file (base64)
      description: Creates parent directories as needed; overwrites existing files.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileWriteRequest" }
      responses:
        "204": { description: Written }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/inspect:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    get:
      tags: [lifecycle]
      summary: Inspect a sandbox
      description: Returns normalized metadata plus the raw SDK config blob.
      responses:
        "200":
          description: Full config + metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Inspect" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/jobs/{job}/stdin:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
      - { name: job, in: path, required: true, schema: { type: string } }
    post:
      tags: [jobs]
      summary: Write to a job's stdin
      description: >
        Writes to the stdin pipe of a job launched with `stdin: true`. Supply
        `data` (UTF-8) or `data_b64` (base64). Set `close_after` to send EOF.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobStdinRequest" }
      responses:
        "204": { description: Written }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/jobs/{job}/signal:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
      - { name: job, in: path, required: true, schema: { type: string } }
    post:
      tags: [jobs]
      summary: Signal a job
      description: Sends a signal to the job's process. `signal <= 0` means kill.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobSignalRequest" }
      responses:
        "204": { description: Signalled }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/list:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: List a directory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FilePathRequest" }
      responses:
        "200":
          description: Directory entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileListResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/stat:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: Stat a path
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FilePathRequest" }
      responses:
        "200":
          description: Path metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileStatResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/exists:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: Check whether a path exists
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FilePathRequest" }
      responses:
        "200":
          description: Existence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileExistsResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/mkdir:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: Create a directory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FilePathRequest" }
      responses:
        "204": { description: Created }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/remove:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: Remove a file or directory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileRemoveRequest" }
      responses:
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sandboxes/{id}/files/copy:
    parameters: [ { $ref: "#/components/parameters/SandboxId" } ]
    post:
      tags: [files]
      summary: Copy a file within the sandbox
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileCopyRequest" }
      responses:
        "204": { description: Copied }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/images/prune:
    post:
      tags: [images]
      summary: Remove image artifacts unused by any sandbox
      responses:
        "200":
          description: Prune report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImagePruneReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A token from `MSBD_API_KEY` or from the persisted key store
        (`msbd keys create`). Stored tokens are shown once at creation and kept
        only as a SHA-256 hash.
...